scripts/requirements-plate.txt
ultralytics
paddleocr
paddlepaddle-gpu
torch
torchvision
numpy
pillow
